+++ TTKit.framework/Versions/A/Headers/TTPresenceData.h
@@ -8,6 +8,8 @@
 
 #import <Foundation/Foundation.h>
 
+@protocol TTPresenceDataObserver;
+
 /**
  *  When set to the type property, marks the respective TTUser as available
  */
@@ -46,16 +48,32 @@ FOUNDATION_EXPORT NSString *const kTTKitPresenceDataTypeUnavailable;
 /**
  * Block that is called when Presence data type is changed
  */
-@property (nonatomic, copy) void (^typeChangedBlock)(NSString*);
+@property (nonatomic, copy) void (^typeChangedBlock)(NSString*) DEPRECATED_MSG_ATTRIBUTE("Use TTPresenceDataObserver instead.");
 
 /**
  * Block that is called when Presence data status is changed
  */
-@property (nonatomic, copy) void (^statusChangedBlock)(NSString*);
+@property (nonatomic, copy) void (^statusChangedBlock)(NSString*) DEPRECATED_MSG_ATTRIBUTE("Use TTPresenceDataObserver instead.");
 
 /**
- * Nilifies status and type presence changed blocks
+ *  Add an observer to track presence and user status changes.
  */
-- (void)prepareForReuse;
+- (void)addObserver:(id<TTPresenceDataObserver>)observer;
+
+/**
+ *  Remove an observer.
+ */
+- (void)removeObserver:(id<TTPresenceDataObserver>)observer;
+
+@end
+
+/**
+ @abstract The `TTPresenceDataObserver` notifies its receiver when there's a change in the user online availability or when the user status changed.
+ */
+@protocol TTPresenceDataObserver <NSObject>
+
+@optional
+- (void)presenseDataChanged:(NSString *)available;
+- (void)userStatusChanged:(NSString *)userStatus;
 
 @end